+++ services/follower-api/src/main/java/com/twitter/entity/Follow.java
@@ -2,6 +2,7 @@ package com.twitter.entity;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.experimental.Accessors;
@@ -35,6 +36,7 @@ import java.util.UUID;
     )
 )
 @Data
+@Builder
 @Accessors(chain = true)
 @NoArgsConstructor
 @AllArgsConstructor